add confirm to delete user commit

+++ frontend/src/screens/UserListScreen.js
@@ -32,7 +32,10 @@ const UserListScreen = ( {history} ) => {
     
     const deleteHandler = (id) => {
         // console.log('Delete Handler action!' + id);
-        dispatch( deleteUser(id) );
+        if(window.confirm('Are you sure')){
+            dispatch( deleteUser(id) );
+        }
+        
     }
     
     return (
